Added caching of fields requiring deep clones, changed Clone handler architecture
